AKAS_ImageViewer_Windows/requirements.desktop.txt
# AKAS Viewer Suite – Desktop dependencies
Pillow>=10.3
send2trash>=1.8.2
psd-tools>=1.9.28
pdf2image>=1.16.3
opencv-python-headless>=4.9.0
# Optional: for advanced PDF rendering if pdf2image not enough
PyMuPDF>=1.23.7
